fix: address PR review — security defaults, file split, async IO, opt-out
Critical:
- Security LLM now enabled by default (PARALLAX_DISABLE_SECURITY_LLM
  defaults to false, not true)
- Remove hasAdminContext prompt injection vector — admin trust must come
  from authenticated trust chain, not string matching the payload
- Remove buildSecuritySafeResult — all paths use heuristic classification

Important:
- Add PARALLAX_ACTION_COMPACTION env var (default on, set =0 to disable)
- Replace synchronous writeFileSync/mkdirSync with async writeFile/mkdir
  in prompt capture path
- Split 625-line file into prompt-optimization.ts (355 LOC) and
  prompt-compaction.ts (276 LOC) — both under 500 LOC limit
- Document PARALLAX_SECURITY_GATING vs DISABLE_SECURITY_LLM interaction
