+++ packages/agent/src/runtime/__tests__/prompt-optimization-actions.test.ts
@@ -16,7 +16,7 @@ import {
   buildFullParamActionSet,
   compactActionsForIntent,
   detectIntentCategories,
-} from "../prompt-optimization";
+} from "../prompt-compaction";
 
 // ---------------------------------------------------------------------------
 // Test fixtures
