+++ packages/agent/src/runtime/prompt-optimization.ts
@@ -6,9 +6,14 @@
  * is controlled via PARALLAX_* env vars (see .env.example).
  */
 
-import { mkdirSync, writeFileSync } from "node:fs";
+import { mkdir, writeFile } from "node:fs/promises";
 import path from "node:path";
 import type { AgentRuntime } from "@elizaos/core";
+
+import { compactActionsForIntent, compactModelPrompt } from "./prompt-compaction";
+
+// Re-export compaction functions for backwards compatibility
+export { detectIntentCategories, buildFullParamActionSet, compactActionsForIntent } from "./prompt-compaction";
 
 // ---------------------------------------------------------------------------
 // Env-var driven configuration (evaluated once at import time)
@@ -26,11 +31,23 @@ const PARALLAX_EMBEDDING_FASTPATH =
   process.env.PARALLAX_EMBEDDING_FASTPATH === "1" ||
   process.env.PARALLAX_EMBEDDING_FASTPATH?.toLowerCase() === "true";
 
+/**
+ * When true, the security evaluation LLM call is completely bypassed and
+ * replaced with a local keyword heuristic.  Defaults to false (security
+ * LLM enabled).  Set PARALLAX_DISABLE_SECURITY_LLM=1 to explicitly
+ * opt-in to disabling the security LLM.
+ *
+ * Interaction with PARALLAX_SECURITY_GATING: when the security LLM is
+ * disabled (this flag = true), gating is irrelevant because no LLM call
+ * is ever made.  When the security LLM is enabled (this flag = false),
+ * PARALLAX_SECURITY_GATING controls whether low-risk messages skip the
+ * LLM call (only high-risk messages are sent to the LLM).
+ */
 const PARALLAX_DISABLE_SECURITY_LLM = (() => {
   const raw = process.env.PARALLAX_DISABLE_SECURITY_LLM?.toLowerCase();
   if (raw === "0" || raw === "false") return false;
   if (raw === "1" || raw === "true") return true;
-  return true;
+  return false;
 })();
 
 const PARALLAX_SECURITY_GATING = (() => {
@@ -49,276 +66,14 @@ const PARALLAX_CAPTURE_PROMPTS =
   process.env.PARALLAX_CAPTURE_PROMPTS === "1" ||
   process.env.PARALLAX_CAPTURE_PROMPTS?.toLowerCase() === "true";
 
+/** When false, context-aware action compaction is skipped entirely. Default: enabled. */
+const PARALLAX_ACTION_COMPACTION = (() => {
+  const raw = process.env.PARALLAX_ACTION_COMPACTION?.toLowerCase();
+  if (raw === "0" || raw === "false") return false;
+  return true;
+})();
+
 let promptCaptureSeq = 0;
-
-// ---------------------------------------------------------------------------
-// Prompt compaction helpers
-// ---------------------------------------------------------------------------
-
-function compactInitialCodeMarker(prompt: string): string {
-  return prompt.replace(
-    /initial code:\s*([0-9a-f]{8})[0-9a-f-]*/gi,
-    "<initial_code>$1</initial_code>",
-  );
-}
-
-// compactActionDocs removed — replaced by compactActionsForIntent which
-// provides context-aware action formatting instead of blanket compaction.
-
-function compactRegistryCatalog(prompt: string): string {
-  return prompt.replace(
-    /\*\*Available Plugins from Registry \((\d+) total\):[\s\S]*?(?=\n## Project Context \(Workspace\)|\n### AGENTS\.md|$)/g,
-    (_match, total: string) =>
-      `**Available Plugins from Registry (${total} total):** [omitted in compact mode; query on demand]\n`,
-  );
-}
-
-function compactCodingActionExamples(prompt: string): string {
-  const next = prompt.replace(
-    /\n# Coding Agent Action Call Examples[\s\S]*?(?=\nPossible response actions:|\n# Available Actions|\n## Project Context \(Workspace\)|$)/g,
-    "\n",
-  );
-  return next.replace(/\nPossible response actions:[^\n]*\n?/g, "\n");
-}
-
-function compactUiCatalog(prompt: string): string {
-  return prompt.replace(
-    /\n## Rich UI Output — you can render interactive components in your replies[\s\S]*?(?=\n## Project Context \(Workspace\)|\n### AGENTS\.md|$)/g,
-    "\n",
-  );
-}
-
-function compactLoadedPluginLists(prompt: string): string {
-  const loadedCountMatch = prompt.match(
-    /\*\*Loaded Plugins:\*\*[\s\S]*?(?=\n\*\*System Plugins:\*\*)/,
-  );
-  const loadedCount = loadedCountMatch
-    ? (loadedCountMatch[0].match(/\n- /g)?.length ?? 0)
-    : 0;
-
-  return prompt.replace(
-    /\n\*\*Loaded Plugins:\*\*[\s\S]*?(?=\n\*\*Available Plugins from Registry|\nNo access to role information|\nSECURITY ALERT:|$)/g,
-    `\n**Loaded Plugins:** ${loadedCount} loaded [list omitted in compact mode]`,
-  );
-}
-
-function compactEmoteCatalog(prompt: string): string {
-  return prompt.replace(
-    /\n## Available Emotes[\s\S]*?(?=\n# Active Workspaces & Agents|\n## Project Context \(Workspace\)|$)/g,
-    "\n## Available Emotes\n[emote catalog omitted in compact mode]\n",
-  );
-}
-
-function compactWorkspaceContextForNonCoding(prompt: string): string {
-  return prompt.replace(
-    /\n## Project Context \(Workspace\)[\s\S]*?(?=\nAdmin trust:|\nThe current date and time is|\n# Conversation Messages|$)/g,
-    "\n## Project Context (Workspace)\n[workspace file contents omitted in compact mode for non-coding intent]\n",
-  );
-}
-
-function compactUiComponentCatalog(prompt: string): string {
-  return prompt.replace(
-    /\n### Available components \((\d+) total\)[\s\S]*?(?=\n## Available Emotes|\n## Project Context \(Workspace\)|$)/g,
-    (_match, total: string) =>
-      `\n### Available components (${total} total)\n[component catalog omitted in compact mode]\n`,
-  );
-}
-
-function compactInstalledSkills(prompt: string): string {
-  return prompt.replace(
-    /\n## Installed Skills \((\d+)\)[\s\S]*?\*Use TOGGLE_SKILL to enable\/disable skills\.[\s\S]*?(?=\nMima is|\n\*\*Loaded Plugins:\*\*|\n## Project Context \(Workspace\)|$)/g,
-    (_match, total: string) =>
-      `\n## Installed Skills (${total})\n[skill list omitted in compact mode; query on demand]\n`,
-  );
-}
-
-// ---------------------------------------------------------------------------
-// Intent detection
-// ---------------------------------------------------------------------------
-
-const CODING_INTENT_RE =
-  /\b(code|coding|repo|repository|pull request|pr\b|branch|test(s)?\b|compile|build|debug|fix|start_coding_task|spawn_coding_agent|send_to_coding_agent)\b|https?:\/\/(?:github\.com|gitlab\.com|bitbucket\.org)\//i;
-const PLUGIN_UI_INTENT_RE =
-  /\b(plugin|plugins|configure|configuration|setup|install|enable|disable|api key|credential|secret|dashboard|form|ui|interface|\[config:)\b/i;
-const TERMINAL_INTENT_RE =
-  /\b(shell|command|execute|run|npm|bun|git\b|bash|terminal|script|pip)\b/i;
-const EMOTE_INTENT_RE =
-  /\b(emote|wave|dance|bow|clap|laugh|angry|sad|think|sit|play_emote)\b/i;
-const ISSUE_INTENT_RE =
-  /\b(issue|bug|ticket|label|close|reopen|github issue|create issue)\b/i;
-
-/** Actions that are always included at full detail. */
-const UNIVERSAL_ACTIONS = new Set(["REPLY", "NONE", "IGNORE"]);
-
-/** Map intent categories → action names that get full params when detected. */
-const INTENT_ACTION_MAP: Record<string, Set<string>> = {
-  coding: new Set([
-    "START_CODING_TASK",
-    "SPAWN_CODING_AGENT",
-    "PROVISION_WORKSPACE",
-    "FINALIZE_WORKSPACE",
-    "LIST_CODING_AGENTS",
-    "SEND_TO_CODING_AGENT",
-  ]),
-  terminal: new Set(["RUN_IN_TERMINAL", "RESTART_AGENT"]),
-  issues: new Set(["MANAGE_ISSUES"]),
-  emote: new Set(["PLAY_EMOTE"]),
-};
-
-function hasIntent(prompt: string, keywords: RegExp): boolean {
-  const taskMatch = prompt.match(/<task>([\s\S]*?)<\/task>/i);
-  const taskText = (taskMatch?.[1] ?? "").slice(0, 2000);
-  if (keywords.test(taskText)) return true;
-
-  // Extract just the user's message line(s) from "# Received Message".
-  // The section also contains instructions with generic words like "execute",
-  // "run", "command" — only match against the actual user text.
-  const msgSection = prompt.indexOf("# Received Message");
-  if (msgSection !== -1) {
-    const afterHeader = prompt.slice(msgSection + "# Received Message".length);
-    // User message is between the header and the next section marker (# or <)
-    const nextSection = afterHeader.search(/\n#|\n<|\n\n\n/);
-    const userMsg = (
-      nextSection !== -1 ? afterHeader.slice(0, nextSection) : afterHeader.slice(0, 500)
-    ).trim();
-    if (keywords.test(userMsg)) return true;
-  }
-
-  // Fallback: scan last few user messages in the conversation
-  const convSection = prompt.indexOf("# Conversation Messages");
-  if (convSection !== -1) {
-    const convBlock = prompt.slice(convSection, prompt.indexOf("# Received Message", convSection));
-    // Match only lines that start with "User:" or "user:"
-    const userLines = convBlock.match(/^(?:User|user):.*$/gm);
-    if (userLines) {
-      const recentUserText = userLines.slice(-3).join(" ");
-      if (keywords.test(recentUserText)) return true;
-    }
-  }
-
-  return false;
-}
-
-/**
- * Detect which intent categories are present in the prompt.
- * Returns array of category names (e.g. ["coding", "terminal"]).
- * Multiple categories can match simultaneously.
- */
-export function detectIntentCategories(prompt: string): string[] {
-  const categories: string[] = [];
-  if (hasIntent(prompt, CODING_INTENT_RE)) categories.push("coding");
-  if (hasIntent(prompt, TERMINAL_INTENT_RE)) categories.push("terminal");
-  if (hasIntent(prompt, ISSUE_INTENT_RE)) categories.push("issues");
-  if (hasIntent(prompt, EMOTE_INTENT_RE)) categories.push("emote");
-  if (hasIntent(prompt, PLUGIN_UI_INTENT_RE)) categories.push("plugin_ui");
-  return categories;
-}
-
-/**
- * Build the set of action names that should get full param detail.
- * Universal actions are always included. Intent-matched actions are
- * added based on detected categories. Everything else gets stub-only.
- */
-export function buildFullParamActionSet(
-  intentCategories: string[],
-): Set<string> {
-  const fullActions = new Set(UNIVERSAL_ACTIONS);
-  for (const cat of intentCategories) {
-    const actions = INTENT_ACTION_MAP[cat];
-    if (actions) {
-      for (const a of actions) fullActions.add(a);
-    }
-  }
-  // Coding intent also implies terminal + issues
-  if (intentCategories.includes("coding")) {
-    for (const a of INTENT_ACTION_MAP.terminal) fullActions.add(a);
-    for (const a of INTENT_ACTION_MAP.issues) fullActions.add(a);
-  }
-  return fullActions;
-}
-
-/**
- * Context-aware action formatting. Replaces the <actions>...</actions>
- * block in the prompt with a version where only intent-relevant actions
- * have full <params> — the rest are stubs with just name + description.
- *
- * If no intents are detected, keeps all params (safe fallback).
- */
-export function compactActionsForIntent(prompt: string): string {
-  // Find the first <actions>...</actions> block (the Available Actions section)
-  const actionsStart = prompt.indexOf("<actions>");
-  if (actionsStart === -1) return prompt;
-  const actionsEnd = prompt.indexOf("</actions>", actionsStart);
-  if (actionsEnd === -1) return prompt;
-
-  const actionsBlock = prompt.slice(
-    actionsStart + "<actions>".length,
-    actionsEnd,
-  );
-
-  const intentCategories = detectIntentCategories(prompt);
-  // When no specific intent is detected, it's general chat — only universal
-  // actions (REPLY, NONE, IGNORE) need full detail. All other actions get
-  // stubs so the LLM knows they exist but doesn't waste context on params.
-  const fullParamActions = buildFullParamActionSet(intentCategories);
-
-  // Parse individual <action>...</action> blocks
-  const actionRegex = /<action>([\s\S]*?)<\/action>/g;
-  const compactedActions: string[] = [];
-
-  for (const match of actionsBlock.matchAll(actionRegex)) {
-    const actionInner = match[1];
-    const nameMatch = actionInner.match(/<name>([\s\S]*?)<\/name>/);
-    if (!nameMatch) continue;
-
-    const actionName = nameMatch[1].trim();
-
-    if (fullParamActions.has(actionName)) {
-      // Keep full action with params
-      compactedActions.push(`  <action>${actionInner}</action>`);
-    } else {
-      // Stub: name + description only, strip <params>
-      const descMatch = actionInner.match(
-        /<description>([\s\S]*?)<\/description>/,
-      );
-      const desc = descMatch?.[1]?.trim() ?? "";
-      compactedActions.push(
-        `  <action>\n    <name>${actionName}</name>\n    <description>${desc}</description>\n  </action>`,
-      );
-    }
-  }
-
-  const compactedBlock = `<actions>\n${compactedActions.join("\n")}\n</actions>`;
-  return `${prompt.slice(0, actionsStart)}${compactedBlock}${prompt.slice(actionsEnd + "</actions>".length)}`;
-}
-
-function compactModelPrompt(prompt: string): string {
-  const hasCodingIntent = hasIntent(prompt, CODING_INTENT_RE);
-  const hasPluginUiIntent = hasIntent(prompt, PLUGIN_UI_INTENT_RE);
-
-  let next = prompt;
-  next = compactInitialCodeMarker(next);
-  if (!hasCodingIntent) {
-    next = compactCodingActionExamples(next);
-  }
-  // Context-aware action compaction (replaces old compactActionDocs)
-  next = compactActionsForIntent(next);
-  next = compactLoadedPluginLists(next);
-  next = compactEmoteCatalog(next);
-  if (!hasCodingIntent) {
-    next = compactInstalledSkills(next);
-  }
-  if (!hasPluginUiIntent) {
-    next = compactRegistryCatalog(next);
-    next = compactUiCatalog(next);
-  } else {
-    next = compactUiComponentCatalog(next);
-  }
-  if (!hasCodingIntent) {
-    next = compactWorkspaceContextForNonCoding(next);
-  }
-  return next;
-}
 
 // ---------------------------------------------------------------------------
 // Security eval helpers
@@ -336,20 +91,6 @@ function isHighRiskMessage(text: string): boolean {
   );
 }
 
-function buildSecuritySafeResult(message: string): string {
-  return JSON.stringify({
-    detected: false,
-    confidence: 0.2,
-    type: "none",
-    severity: "low",
-    reasoning:
-      message.length > 0
-        ? "Fast-path low-risk classification in compact mode."
-        : "Fast-path low-risk classification in compact mode (no analyzable message).",
-    indicators: [],
-  });
-}
-
 function buildSecurityHeuristicResult(message: string): string {
   const highRisk = isHighRiskMessage(message);
   return JSON.stringify({
@@ -362,12 +103,6 @@ function buildSecurityHeuristicResult(message: string): string {
       : "Local heuristic classified message as low-risk.",
     indicators: highRisk ? ["keyword_match"] : [],
   });
-}
-
-function hasAdminContext(prompt: string): boolean {
-  return /\b(admin trust: current speaker is world owner|current speaker is world owner|role:\s*admin|speaker.*\bowner\b|world owner)\b/i.test(
-    prompt,
-  );
 }
 
 // ---------------------------------------------------------------------------
@@ -467,10 +202,10 @@ export function installPromptOptimizations(runtime: AgentRuntime): void {
     if (PARALLAX_CAPTURE_PROMPTS) {
       try {
         const captureDir = path.resolve(".tmp", "prompt-captures");
-        mkdirSync(captureDir, { recursive: true });
+        await mkdir(captureDir, { recursive: true });
         const seq = String(++promptCaptureSeq).padStart(4, "0");
         const filename = `${seq}-${modelType}.txt`;
-        writeFileSync(
+        await writeFile(
           path.join(captureDir, filename),
           `--- model: ${modelType} | key: ${promptKey} | chars: ${originalPrompt.length} ---\n\n${originalPrompt}`,
         );
@@ -486,7 +221,6 @@ export function installPromptOptimizations(runtime: AgentRuntime): void {
       originalPrompt.startsWith("You are a security evaluation system.")
     ) {
       const analyzedMessage = extractSecurityMessage(originalPrompt);
-      const adminContext = hasAdminContext(originalPrompt);
       const cacheKey = analyzedMessage.slice(0, 1000);
       const now = Date.now();
       const cacheTtlMs = 5 * 60_000;
@@ -498,15 +232,11 @@ export function installPromptOptimizations(runtime: AgentRuntime): void {
         }
         return cached.value;
       }
-      const heuristic = adminContext
-        ? buildSecuritySafeResult(analyzedMessage)
-        : buildSecurityHeuristicResult(analyzedMessage);
+      const heuristic = buildSecurityHeuristicResult(analyzedMessage);
       rt.__miladySecurityEvalCache.set(cacheKey, { at: now, value: heuristic });
       if (PARALLAX_PROMPT_TRACE) {
         runtime.logger?.info(
-          adminContext
-            ? "[milady] Security LLM disabled; admin context fast-pathed safe"
-            : "[milady] Security LLM disabled; using local heuristic",
+          "[milady] Security LLM disabled; using local heuristic",
         );
       }
       return heuristic;
@@ -534,14 +264,14 @@ export function installPromptOptimizations(runtime: AgentRuntime): void {
           return cached.value;
         }
         if (!isHighRisk) {
-          const safe = buildSecuritySafeResult(analyzedMessage);
-          rt.__miladySecurityEvalCache.set(cacheKey, { at: now, value: safe });
+          const heuristic = buildSecurityHeuristicResult(analyzedMessage);
+          rt.__miladySecurityEvalCache.set(cacheKey, { at: now, value: heuristic });
           if (PARALLAX_PROMPT_TRACE) {
             runtime.logger?.info(
               "[milady] Security eval skipped for low-risk prompt",
             );
           }
-          return safe;
+          return heuristic;
         }
       }
 
@@ -569,10 +299,10 @@ export function installPromptOptimizations(runtime: AgentRuntime): void {
       }
     }
 
-    // --- Context-aware action compaction (always active) ---
+    // --- Context-aware action compaction (when enabled) ---
     // Strips <params> from actions not relevant to the user's intent.
     // Safe to run always: all action names remain visible, only detail is stripped.
-    let workingPrompt = isTextLarge
+    let workingPrompt = isTextLarge && PARALLAX_ACTION_COMPACTION
       ? compactActionsForIntent(originalPrompt)
       : originalPrompt;
